+++ metadata/apps/example-hasura/router.demo.js
@@ -4,6 +4,9 @@ import asyncHandler from 'express-async-handler';
 // Import our custom request logger
 import hasuraRequestLogger from './hasura-request-logger.js';
 
+// Import protector for web hooks
+import hasuraWebHookProtector from './protector.js';
+
 // Import our errors definition
 import errors from './errors.demo.js';
 
@@ -12,10 +15,21 @@ const router = express.Router();
 
 router.use('/hasura-*', hasuraRequestLogger);
 
+router.use('/hasura-*', hasuraWebHookProtector);
+
 router.post(
   '/hasura-version',
   asyncHandler((req, res) => {
     console.log(`${process.env.APP_VERSION}`);
+    res.json({
+      version: `${process.env.APP_VERSION}`,
+    });
+  })
+);
+
+router.post(
+  '/hasura-version-error',
+  asyncHandler((req, res) => {
     throw new Error(errors.SOMETHING_WENT_TERRIBLY_WRONG);
     // res.json({
     //   version: `${process.env.APP_VERSION}`,